+++ src/main.rs
@@ -3,22 +3,23 @@ use tref;
 
 fn main() {
     if let Ok(file) = File::open("file.tref") {
-        let model = <tref::Model>::new();
-        match model.parse(BufReader::new(file)) {
+        match <tref::Model>::parse(BufReader::new(file)) {
             Ok(forest) => {
-                if let Some(tree) = forest.get_tree("my_tree") {
-                    println!("my_tree = {:#?}", tree);
-                }
-
-                let f = File::create("serialized.tref").expect("Unable to create file");
-                let mut buf_writer = BufWriter::new(f);
-                match model.serialize(&forest, &mut buf_writer) {
-                    Ok(num_lines) => {
-                        println!("Tree serialized correctly, num lines = {}", num_lines);
-                    },
-                    Err(e) => {
-                        println!("Failed serializing tree: {}", e);
+                for (tree_id, tree) in forest.iter() {
+                    println!("{} = {:#?}", tree_id, tree);
+    
+                    let f = File::create("serialized.tref").expect("Unable to create file");
+                    let mut buf_writer = BufWriter::new(f);
+                    match <tref::Model>::serialize(&forest, &mut buf_writer) {
+                        Ok(num_lines) => {
+                            println!("Tree serialized correctly, num lines = {}", num_lines);
+                        },
+                        Err(e) => {
+                            println!("Failed serializing tree: {}", e);
+                        }
                     }
+
+                    println!("--------------------------");
                 }
             },
             Err(e) => {